refactor: refactor List Item component
