+++ types/Product.ts
@@ -1,0 +1,8 @@
+export interface Product {
+    name: string
+    description: string
+    dateClose: string
+    dateOpen: string
+    link: string
+    type: "app" | "hardware" | "service"
+}
